Exception messages

+++ src/main/java/com/brasileirao/championship/Services/TeamsServices.java
@@ -5,6 +5,7 @@ import com.brasileirao.championship.Repositories.TeamsRepository;
 import com.brasileirao.championship.Services.Exceptions.ResourceNotFoundException;
 import com.brasileirao.championship.dto.TeamsDto;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -25,33 +26,30 @@ public class TeamsServices {
     @Transactional(readOnly = true)
     public TeamsDto findTeamById(Long id) {
         Optional<Teams> obj = repository.findById(id);
-        Teams teams = obj.orElseThrow(() -> new ResourceNotFoundException("Team not found !"));
+        Teams teams = obj.orElseThrow(() -> new ResourceNotFoundException("Id not found: " + id));
         return new TeamsDto(teams);
     }
 
     @Transactional(readOnly = true)
     public TeamsDto findAllByTeam(String team) {
         Optional<Teams> obj = repository.findAllByTeam(team);
-        Teams teamObj = obj.orElseThrow(() -> new ResourceNotFoundException("Team not found!"));
+        Teams teamObj = obj.orElseThrow(() -> new ResourceNotFoundException("Team not found: " + team));
         return new TeamsDto(teamObj);
     }
 
     @Transactional
     public TeamsDto updateTeams(Long id, TeamsDto teams) {
-        try {
-            Teams entity = repository.getReferenceById(id);
-            entity.setTeam(teams.getTeam());
-            entity.setCity(teams.getCity());
-            entity.setFullName(teams.getFullName());
-            entity.setAcronym(teams.getAcronym());
-            entity.setRegion(teams.getRegion());
-            entity.setState(teams.getState());
-            entity.setStadium(teams.getStadium());
-            entity = repository.save(entity);
-            return new TeamsDto(entity);
-        } catch (Exception e) {
-            throw new RuntimeException(e);
-        }
+        String team = teams.getTeam();
+        Teams entity = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Team not found with id: " + id + " - " + team));
+        entity.setTeam(teams.getTeam());
+        entity.setCity(teams.getCity());
+        entity.setFullName(teams.getFullName());
+        entity.setAcronym(teams.getAcronym());
+        entity.setRegion(teams.getRegion());
+        entity.setState(teams.getState());
+        entity.setStadium(teams.getStadium());
+        entity = repository.save(entity);
+        return new TeamsDto(entity);
     }
 
     @Transactional(readOnly = true)
@@ -72,10 +70,11 @@ public class TeamsServices {
     }
 
     public void deleteTeam(Long id) {
-        try {
-            repository.deleteById(id);
-        } catch (Exception e) {
-            throw new RuntimeException(e);
+        boolean existsTeam = repository.existsById(id);
+        if (!existsTeam) {
+            throw new ResourceNotFoundException("Id not found: " + id);
         }
+        repository.deleteById(id);
     }
 }
+
